@@ -1,36 +1,94 @@
 {
-  "name": "new",
-  "version": "0.1.0",
-  "private": true,
+  "name": "logankuzyk-com",
+  "version": "1.0.0",
+  "description": "Website template for Payload",
+  "license": "MIT",
+  "type": "module",
   "scripts": {
-    "dev": "next dev",
-    "build": "next build",
-    "start": "next start",
-    "lint": "next lint",
-    "prepare": "husky install"
+    "build": "cross-env NODE_OPTIONS=--no-deprecation next build",
+    "postbuild": "next-sitemap --config next-sitemap.config.cjs",
+    "dev": "cross-env NODE_OPTIONS=--no-deprecation next dev",
+    "dev:prod": "cross-env NODE_OPTIONS=--no-deprecation rm -rf .next && pnpm build && pnpm start",
+    "generate:importmap": "cross-env NODE_OPTIONS=--no-deprecation payload generate:importmap",
+    "generate:types": "cross-env NODE_OPTIONS=--no-deprecation payload generate:types",
+    "ii": "cross-env NODE_OPTIONS=--no-deprecation pnpm --ignore-workspace install",
+    "lint": "cross-env NODE_OPTIONS=--no-deprecation next lint",
+    "lint:fix": "cross-env NODE_OPTIONS=--no-deprecation next lint --fix",
+    "payload": "cross-env NODE_OPTIONS=--no-deprecation payload",
+    "reinstall": "cross-env NODE_OPTIONS=--no-deprecation rm -rf node_modules && rm pnpm-lock.yaml && pnpm --ignore-workspace install",
+    "start": "cross-env NODE_OPTIONS=--no-deprecation next start",
+    "test": "pnpm run test:int && pnpm run test:e2e",
+    "test:e2e": "cross-env NODE_OPTIONS=\"--no-deprecation --no-experimental-strip-types\" pnpm exec playwright test --config=playwright.config.ts",
+    "test:int": "cross-env NODE_OPTIONS=--no-deprecation vitest run --config ./vitest.config.mts"
   },
   "dependencies": {
-    "@types/node": "20.6.0",
-    "@types/react": "18.2.21",
-    "@types/react-dom": "18.2.7",
-    "autoprefixer": "^10.4.15",
-    "classnames": "^2.3.2",
-    "eslint": "8.49.0",
-    "eslint-config-next": "13.4.19",
-    "eslint-plugin-import": "^2.28.1",
-    "firebase": "^10.3.1",
-    "flowbite": "^1.8.1",
-    "flowbite-react": "^0.5.0",
-    "husky": "^8.0.3",
-    "next": "13.4.19",
-    "postcss": "^8.4.29",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
-    "react-feather": "^2.0.10",
-    "tailwindcss": "^3.3.3",
-    "typescript": "5.2.2"
+    "@payloadcms/admin-bar": "3.49.1",
+    "@payloadcms/live-preview-react": "3.49.1",
+    "@payloadcms/next": "3.49.1",
+    "@payloadcms/payload-cloud": "3.49.1",
+    "@payloadcms/plugin-form-builder": "3.49.1",
+    "@payloadcms/plugin-nested-docs": "3.49.1",
+    "@payloadcms/plugin-redirects": "3.49.1",
+    "@payloadcms/plugin-search": "3.49.1",
+    "@payloadcms/plugin-seo": "3.49.1",
+    "@payloadcms/richtext-lexical": "3.49.1",
+    "@payloadcms/ui": "3.49.1",
+    "@radix-ui/react-checkbox": "^1.0.4",
+    "@radix-ui/react-label": "^2.0.2",
+    "@radix-ui/react-select": "^2.0.0",
+    "@radix-ui/react-slot": "^1.0.2",
+    "class-variance-authority": "^0.7.0",
+    "clsx": "^2.1.1",
+    "cross-env": "^7.0.3",
+    "dotenv": "16.4.7",
+    "geist": "^1.3.0",
+    "graphql": "^16.8.2",
+    "lucide-react": "^0.378.0",
+    "next": "15.4.4",
+    "next-sitemap": "^4.2.3",
+    "payload": "3.49.1",
+    "prism-react-renderer": "^2.3.1",
+    "react": "19.1.0",
+    "react-dom": "19.1.0",
+    "react-hook-form": "7.45.4",
+    "sharp": "0.34.2",
+    "tailwind-merge": "^2.3.0",
+    "tailwindcss-animate": "^1.0.7",
+    "@payloadcms/db-postgres": "3.49.1"
   },
   "devDependencies": {
-    "husky": "^8.0.0"
+    "@eslint/eslintrc": "^3.2.0",
+    "@playwright/test": "1.54.1",
+    "@tailwindcss/typography": "^0.5.13",
+    "@testing-library/react": "16.3.0",
+    "@types/escape-html": "^1.0.2",
+    "@types/node": "22.5.4",
+    "@types/react": "19.1.8",
+    "@types/react-dom": "19.1.6",
+    "@vitejs/plugin-react": "4.5.2",
+    "autoprefixer": "^10.4.19",
+    "copyfiles": "^2.4.1",
+    "eslint": "^9.16.0",
+    "eslint-config-next": "15.4.4",
+    "jsdom": "26.1.0",
+    "playwright": "1.54.1",
+    "playwright-core": "1.54.1",
+    "postcss": "^8.4.38",
+    "prettier": "^3.4.2",
+    "tailwindcss": "^3.4.3",
+    "typescript": "5.7.3",
+    "vite-tsconfig-paths": "5.1.4",
+    "vitest": "3.2.3"
+  },
+  "engines": {
+    "node": "^18.20.2 || >=20.9.0",
+    "pnpm": "^9 || ^10"
+  },
+  "pnpm": {
+    "onlyBuiltDependencies": [
+      "sharp",
+      "esbuild",
+      "unrs-resolver"
+    ]
   }
 }
